requirements.txt
alembic==1.13.1
bcrypt==4.1.3
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7

Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5

PyJWT==2.8.0
SQLAlchemy==2.0.30
typing_extensions==4.12.1
Werkzeug==3.0.3

mysqlclient==2.2.4
Flask-MySQLdb==2.0.0

adal==1.2.7
cachetools==4.2.4
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1

crayons==0.4.0
cryptography==39.0.1
DateTime==4.3
distlib==0.3.4
dropbox==11.36.0
ephem==4.1.4
et-xmlfile==1.0.1
filelock==3.6.0
google-api-core==1.31.0
google-api-python-client==2.12.0
google-auth==1.32.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0

httplib2==0.21.0
idna==3.4
jdcal==1.4.1
LunarCalendar==0.0.9
mysql-client==0.0.1
mysql-connector==2.2.9
mysql-connector-python==8.0.17
numpy
oauthlib==3.2.2
openpyxl==3.0.5
packaging==22.0
pandas==1.0.3
Pillow==8.0.1
platformdirs==2.5.2
ply==3.11
protobuf==3.19.1
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyodbc==4.0.30
pyparsing==3.0.9
python-dateutil==2.8.1
pytz==2020.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.14.0
SQLAlchemy==1.3.22
stone==3.3.1
uritemplate==3.0.1
urllib3==1.26.13
virtualenv==20.14.1
xlrd==2.0.1
zope.interface==5.1.0
